requirements.txt
GitPython>=3.1.0
python-dotenv>=1.0.0

